requirements.txt
discord.py
langchain
transformers 
chromadb
huggingface_hub
torch
sentence-transformers